requirements.txt
azure-kusto-data
azure-kusto-ingest
python-dotenv
yfinance